feat : 답변자와 질문자가 다르면 질문 글 삭제 불가 구현
질문글과 답변자가 다르면 삭제 불가하다

+++ src/main/java/com/codessquad/qna/service/QuestionService.java
@@ -44,7 +44,10 @@ public class QuestionService {
 
     @Transactional
     public void delete(Long id) {
-        questionRepository.deleteById(id);
+        Question question = findById(id);
+        if (question.checkSameUserFromAnswer()) {
+            questionRepository.deleteById(id);
+        }
     }
 
 }
